start cleaning up config options for ni card
also split out gated ni card class
